requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
requests==2.32.3
paho-mqtt>=2.0.0
zeroconf>=0.132.0
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.12

# Voice (offline STT + TTS)
vosk==0.3.45
sounddevice==0.4.6
pyttsx3==2.90

# Google Calendar
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1

# Spotify
spotipy==2.24.0

# Tests
pytest==8.3.2
pydantic-settings>=2.5.0
httpx>=0.27
